requirements.txt
mysql-connector-python
pymysql
pandas
tabulate
pymysql
sqlalchemy
